+++ examples/functions/median/function.tf
@@ -1,0 +1,7 @@
+# Return the median average of the element(s) of an even length list:
+provider::stdlib::median([1.2, 5.4, 3.1, 8.9])
+# result => 4.25
+
+# Return the median average of the element(s) of an odd length list:
+provider::stdlib::median([7, 2, 5])
+# result => 5.0